[FEATURE] Restrict visibility of Category for a BE User
The change set is a first step for improving the Category
restriction for a BE User.

The patch brings an additional field in BE User Group / User
where category mount points can be defined.

A security aspect is also introduced for controlling what
mount points are going to be displayed in the tree
widget in TCEforms.

Security check related to Data Handler will be provided in
another patch.

Change-Id: Ie817e03f21ecbc8423e4d4c2c081bb99324be2d2
Releases: 6.2
Resolves: #52718
Reviewed-on: https://review.typo3.org/24618
Reviewed-by: Steffen Ritter
Tested-by: Steffen Ritter