[BUGFIX] Wrong calculation of maximum value for checkbox fields
This patch corrects the calculation of the maximum value for a group
of checkboxes which is stored as bit flag value in the database. The
formular for the maximum value is 2nd power of the item count minus one.

Resolves: #52104
Releases: 6.2, 6.1, 6.0, 4.7, 4.5
Change-Id: I0eb430b72a072838c6ac3bc3f5e339ff2509c455
Reviewed-on: https://review.typo3.org/24653
Reviewed-by: Nicole Cordes
Tested-by: Nicole Cordes
(cherry picked from commit 33d441555e8e9ad8f6c6c6a5157397815597d057)
Reviewed-on: https://review.typo3.org/24654