[FEATURE] Improve usability for BE database thumbnails/file icons
The backend produces thumbnails/file icons/truetype font previews which
come from a database row. The usability for these images can be improved
by showing the file name or title on hover using the title attribute.
Currently only the thumbnail/icon is produced and more information can
only be read by clicking the icon or opening the record, like in the
content element File Links.

This patch changes the alt attribute to the alternative text and adds the
title attribute for these images. When no title has been given for a file
record, the file name will be used instead.

Change-Id: I23b761fb204dda2c200f3ea57961cc017840fcac
Fixes: #47709
Releases: 6.2
Reviewed-on: https://review.typo3.org/20456
Reviewed-by: Alexander Opitz
Tested-by: Alexander Opitz
Reviewed-by: Stefan Neufeind
Tested-by: Stefan Neufeind