[BUGFIX] Missing ENABLE_INSTALL_TOOLs permissions let us loop
If we can't touch ENABLE_INSTALL_TOOL or recreate it as we have not
enough permissions in typo3conf/ we will have an endless loop.

Fixing by not calling createInstallToolEnableFile(), the user must
update the file manually or fix the permissions which are shown as
error in the installer.

Resolves: #52744
Releases: 6.2
Change-Id: Ib3ad316b90f1971dc6b71a35d15a06783f927f7a
Reviewed-on: https://review.typo3.org/24672
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters