[FEATURE] Enable frontend to show translated file metadata
The frontend should show translated file metadata.
Since translation is not the concern of a file abstraction
layer and the FAL does not know about frontend language
settings this concern has to be decoupled.

For that purpose the MetaDataRepository has been extended
with an signal which is fired before the metadata-record
is returned. In Frontend-Context a Slot is registered
which passes that record through the according functionality
in \TYPO3\CMS\Frontend\Page\PageRepository::getRecordOverlay
so that a clean language and workspace overlay happen
according to the settings of TSFE.

Releases: 6.2
Resolves: #52818
Change-Id: I7b7791abd63d2a69480ed31861c63de1585a569c
Reviewed-on: https://review.typo3.org/24766
Reviewed-by: Frans Saris
Tested-by: Frans Saris
Reviewed-by: Fabien Udriot
Reviewed-by: Francois Suter
Tested-by: Francois Suter
Reviewed-by: Ernesto Baschny
Tested-by: Ernesto Baschny