[BUGFIX] Preserve vendor name in refering request
Fluid FormViewHelper adds some hidden fields that contain information
about the current controller. When the form is submitted and contains
a validation error the errorAction int the ActionController is hit.
This one builds a refering request object from the hidden fields without
a vendor name, as this one is not passed from the FormViewHelper.
The request object tries to guess the controller name but guesses a non
namespaced classes name as the vendor name is not given. To get further
information about the class methods the class name gets autoloaded and
creates a wrong classes cache entry as the autoloader cannot detect if
a extension has namespaced or non namespaced classes from the outside.

This patch introduces a hidden vendor name field in the Fluid FormViewHelper.

Fixes: #52823
Releases: 6.2, 6.1, 6.0
Change-Id: I0a82cf2ee07ce293eda0b9f50d3cac7b2a513f15
Reviewed-on: https://review.typo3.org/24788
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters