[TASK] Exclude central Modernizr from concatenation
Modernizr is added to every "page" in the backend
as a default in DocumentTemplate. By default all JS
gets merged. Thus Modernizr is added to multiple
files.

Reduce the size of JS-files and allow browsers to
take advantage of reusing that code by having this
library as a separate part.

Change-Id: If18ca423ac3b4b1347c1319189d0feb019f37b7f
Resolves: #52931
Releases: 6.2, 6.1, 6.0
Reviewed-on: https://review.typo3.org/24878
Reviewed-by: Markus Klein
Tested-by: Markus Klein
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters