[BUGFIX] Fix wrong composer type in filemetadata
The composer file of the filemetadata extension contains the
composer type typo3cms-framework instead of typo3-cms-framework.
This disables the package key translation of the package manager
and the extension is listed internally under the key
typo3.cms.filemetadata. The package manager can handle that key
and its alias filemetadata transparently but the extensionmanager
fails to do so because of its custom approach to installed extension
detection.

This patch fixes the type in the filemetadata composer file.

Fixes: #52925
Releases: 6.2
Change-Id: Id761ff9f9b169f8b559332dd1d54389e9582edf3
Reviewed-on: https://review.typo3.org/24889
Reviewed-by: Francois Suter
Tested-by: Francois Suter
Reviewed-by: Sebastian Fischer
Reviewed-by: Stefan Froemken
Tested-by: Stefan Froemken
Reviewed-by: Steffen Ritter
Tested-by: Steffen Ritter