[BUGFIX] Endless loop in check for broken extensions
The check for broken extensions in the Install Tool checks for
PHP fatal errors, excludes possibly faulty extensions and
continues this process until all errors are gone or all
extensions have been marked as broken.

If PHP warnings (or any other content) are outputted directly
to the client, this will lead to a mixed data representation of
JSON encoded content and the plain text output, like e.g.
  Warning:  array_keys() expects parameter 1 to be array...
  "OK"

To avoid that, the output buffer needs to be cleaned.

Change-Id: I7efb0be23e1897af4d98c7c7726eec2003fb6296
Fixes: #53236
Releases: 6.2
Reviewed-on: https://review.typo3.org/25152
Reviewed-by: Philipp Gampe
Reviewed-by: Markus Klein
Reviewed-by: Oliver Hader
Tested-by: Oliver Hader