[!!!][DB][BUGFIX] Update all storages in upgrade wizard
With commit I805085948e01544efd692498f5e3537612c62050 a new hash was
introduced to sys_file table. The updater misses to update all storages
except the one with uid=1. Furthermore the default storage zero, where
all core related files are stored, is missing completely.
This triggers a fatal error all over the backend.

This patch adds the default storage and fixes the bug, which
did not update other storages beside uid=1.

It is very important to run the upgrade wizard in the
Install Tool after this patch is applied.

Resolves: #53575
Releases: 6.2
Change-Id: Ifa60e8b4e3bf50159b7475cc9af5dab0274340cd
Reviewed-on: https://review.typo3.org/25339
Reviewed-by: Nicole Cordes
Tested-by: Nicole Cordes
Tested-by: Christian Kuhn
Reviewed-by: Christian Kuhn
Reviewed-by: Markus Klein
Tested-by: Markus Klein