[TASK] ext:adodb Remove connection wizard
Extension adodb can be used together with TER extension "datasource".
ext:adodb delivers a wizard and flexform code to connect to this
third party extension. Last release of ext:datasource was in 2005,
the extension in unmaintained and always had an experimental
character. Since core extension functionality should not depend on
third party code, the datasource related code of ext:adodb is fully
removed with this patch in 6.2.

Change-Id: Ib7de137599d9bb55d6b9dd98667cbbe3dd70c986
Resolves: #42651
Releases: 6.2
Reviewed-on: https://review.typo3.org/25759
Reviewed-by: Oliver Klee
Tested-by: Oliver Klee
Reviewed-by: Philipp Gampe
Tested-by: Philipp Gampe
Reviewed-by: Franz G. Jahn
Tested-by: Franz G. Jahn
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn