[SECURITY] feuser_adminLib.inc allows to set arbitrary fields
The CMS core ships a utility class helping extension authors
to create frontend-extension which need a mail-based opt-in.
This class is neither used by core nor really maintained.

In the opt-in process the fields which should be updated to
activate the user are put as URL parameter into the
activation link. In the default configuration this feature
set allows to set any values of any field to this record.

As a result a user could manipulate his activation link and
therefore extend his usergroups.

This patch ensures that all fields which are about to update
are added to the hash as well as only taking the values
from TypoScript so even if the fields match no harm can be
done.

Change-Id: Ic2325cbe714297e6d83e6a6d7436abe4981e8cb1
Fixes: #48187
Releases: 6.0, 4.7, 4.5
Security-Commit: 8bdaffd51a244657215c1d37150f749cbb7a8c79
Security-Bulletin: TYPO3-CORE-SA-2013-004
Reviewed-on: https://review.typo3.org/26188
Reviewed-by: Oliver Hader
Tested-by: Oliver Hader