[!!!][TASK] Driver API has too many crosscutting concerns
The Driver within the FileAbstractionLayer is the actual
abstraction of a concrete file-system meant to unify
access to several kind of storage-systems the file data
might reside in. For each way of storing files there
has to be a driver.

Therefore a driver is solely a unique API to interact with
an underlying storage system based upon its own identifier
system.

Opposite to that, the AbstractDriver (defining our Interface)
knows about its storage, file- and folder-objects as well
as indexes. The methods within the current driver need to
implement more FAL-internal logic than one would expect from
what the driver is purposed to. As a result only a few drivers
are existing by now. In addition it's nearly impossible to
change FAL interals without touching every driver out there.

This change deals with these problems. In several steps the
current driver API has been refactored to have less dependencies
and only fullfills it's original purpose: the abstraction of a
file storage system.

The following things are considered:
   1. Change every return value which has been an object to the
      identifier and adapt the ResourceStorage accordingly.
   2. Change every method parameter being FileInterface, File,
      AbstractFile or Folder objects to a string: the identifier.
   3. Get rid of the storage as a member variable, only the uid
      of the storage is needed at two places.
   4. Remove all abstract methods from the AbstractDriver and
      put them in a newly created DriverInterface.
   5. Exchange all references to AbstractDriver to the interface.
   6. Remove unused methods, which came to light doing so.
   7. Merge addFile/removeFile and *Raw since the do the same if
      no objects are in place.
   8. Unify the creation of folder objects and how information
      is retrieved (analogue to files now).
   9. Unify and simplify the methods for retrieving folder
      contents (files / folders).
  10. Adapt the UnitTests of both classes to the changes.
      This also leads to the removal of some tests which are
      superfluous with that change.
  11. Make the DriverRegistry check the implementation of the
      newly created interface.
  12. Unify the parameter names in the methods.

Releases: 6.2
Resolves: #54230
Resolves: #54231
Change-Id: I4a51572c6a10859e6fd5f12f6cb539950444992e
Reviewed-on: https://review.typo3.org/26520
Reviewed-by: Frans Saris
Tested-by: Frans Saris
Reviewed-by: Benjamin Mack
Tested-by: Benjamin Mack