[BUGFIX] Add stdWrap on value property of TEXT
As discussed back in 2010, the value property of the TEXT content object
should have stdWrap functionality as well, if the HTML content object
is deprecated.

Today the 6.0 TSref already mentions stdWrap for value and
the HTML cObject has been removed with 6.0 as well.

This fix now finally adds the stdWrap.

Resolves: #54371
Releases: 6.2, 6.1, 6.0
Change-Id: I1b7068b3715bc9f8aa4608d5f746c16b70562f3b
Reviewed-on: https://review.typo3.org/26445
Reviewed-by: Benjamin Mack
Reviewed-by: Chris topher
Tested-by: Chris topher
Reviewed-by: Stefan Neufeind
Tested-by: Stefan Neufeind