[BUGFIX] File browser fails on inexistent expandFolder
If a user browsed to a directory "foo" and re-opens the file-
browser later, the same directory is shown expanded.

In case the directory has been moved/renamed/deleted in the
meantime, the user gets an exception due to the inexistent
directory.

This change just adds a try/catch block around to handle that
situation.

Change-Id: I39aab6be46aec1d3f1f365e5d5f5455aca9b3aa2
Resolves: #50266
Releases: 6.2, 6.1, 6.0
Reviewed-on: https://review.typo3.org/26883
Reviewed-by: Markus Klein
Tested-by: Markus Klein