[TASK] Setup travis notifications
Since the travis ci service changed the notification
policy to not notify authors of commits on failed
builds, we need to set up explicit notifications
so that failing builds do not stay undiscovered.

A first step will be to sent a notification to
the irc channel and the core list.

Change-Id: I6f5953f65aab04c526eae8aa34890e39169ab60f
Reviewed-on: https://review.typo3.org/26491
Tested-by: Ernesto Baschny
Reviewed-by: Helmut Hummel
Tested-by: Helmut Hummel