[BUGFIX] PropertyMapper does not work with class aliasses
Currently the property mapper uses the classnames found
in the the class property annotations to find the right
TypeConverter. But for class aliases this breaks as the
alias is used and not the original classname.

There was already a getClassNameForAlias call but this was
to early.

This change moves the getClassNameForAlias() call from
doMapping() to findTypeConverter().

Releases: 6.2, 6.1
Resolves: #54289
Change-Id: Ie4ab3deb3c192290db62ea06826870d2fadba8a7
Reviewed-on: https://review.typo3.org/26578
Reviewed-by: Stefan Froemken
Tested-by: Stefan Froemken
Reviewed-by: Helmut Hummel
Tested-by: Helmut Hummel