[BUGFIX] Replace reference to Security Cookbook with Security Guide
Section "TYPO3 Security" refers to the TYPO3 Security Cookbook, which is
obsolete and has been replaced by the official TYPO3 Security Guide in
December 2011. This patch removes the reference to the Cookbook and
replaces it by a link to the latest version of the Security Guide on
docs.typo3.org.

Resolves: #54756
Releases: 6.2
Change-Id: I4067d40e854e7d4b4f77e2c12239e3e00616f5a7
Reviewed-on: https://review.typo3.org/26656
Reviewed-by: Michiel Roos
Tested-by: Michiel Roos
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters