[TASK] Be more expressive in error condition
If lockSSL is set to 1 and the backend is not
accessed via SSL currently a 404 header is sent
with no message body. The same happens when the
client IP address does not mach the configured
list of allowed addresses.

Give a more useful hint for users in these cases
by throwing an Exception.

Resolves: #51378
Releases: 6.2
Change-Id: Ic76d8415c482e78bdea7e57e1cb5dab21a49969f
Reviewed-on: https://review.typo3.org/26718
Reviewed-by: Michiel Roos
Tested-by: Michiel Roos
Tested-by: Georg Ringer
Reviewed-by: Oliver Klee
Reviewed-by: Markus Klein
Tested-by: Markus Klein
Reviewed-by: Marcin Sągol
Reviewed-by: Felix Kopp
Reviewed-by: Helmut Hummel
Tested-by: Helmut Hummel