[BUGFIX] Edits disappear when published twice
When accidentally publishing an edit twice, eg. by using multiple
browser tabs, all modifications are reverted and the changes are no
longer visible in the workspace.

In the database, the old workspace version of the record has wsid = 0
and pid = -1 and is thus invisible, but still contains the modified
content.

Resolves: #54924
Releases: master, 6.2
Change-Id: I8805737540f030ae24331d33590b71629c7f3179
Reviewed-on: http://review.typo3.org/42592
Reviewed-by: Andreas Wolf <andreas.wolf@typo3.org>
Tested-by: Andreas Wolf <andreas.wolf@typo3.org>