[TASK] Add index to sys_file_metadata
Since sys_file_metadata records are enabled to be 
translated the file list module is capable of metadata 
translations, too. In order to directly link to edit 
or create a new translation depending on wether this 
language already has been translated for each file the 
metadata translations are fetched from the database.

In order to improve that query this patch adds an index 
including l10n_parent and sys_language_uid.

Resolves: #54927
Releases: 6.2
Change-Id: I6cc5c74505e96647c5ddd7e7ee77518176d2ba0b
Reviewed-on: https://review.typo3.org/26774
Reviewed-by: Steffen Ritter
Tested-by: Steffen Ritter
Reviewed-by: Benjamin Mack
Tested-by: Benjamin Mack