[BUGFIX] Display detected fatal in extension check
In the installer fatals can happen during updates of
very outdated extensions. Many updates can lead to a
redirect to the extension check where the user gets
displayed a message that a fatal has been detected.
But the detected fatal error itself is not displayed.

This patch passes the error as url parameter to the
extension check and displays it there.

Fixes: #54943
Releases: 6.2
Change-Id: I111df5d5411015b21c4b2b5e9be3b83b311c8bc5
Reviewed-on: https://review.typo3.org/26781
Reviewed-by: Helmut Hummel
Tested-by: Helmut Hummel
Reviewed-by: Thomas Maroschik
Reviewed-by: Wouter Wolters
Reviewed-by: Stefan Neufeind
Reviewed-by: Markus Klein
Reviewed-by: Anja Leichsenring
Tested-by: Anja Leichsenring
Tested-by: Wouter Wolters
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn