[TASK] Speedup typolink root-line handling
The link generation via typolink needs the domain record as well
as the rootline to generate the links correctly. To save
superfluous database queries some local caches are introduced.

Furthermore the rootline cache does not use all information which
is present after the database query. This information is unpacked
here, too. As a result less requests to the CacheFrontend need to
be done.

Resolves: #54959
Releases: 6.2, 6.1
Change-Id: Iea7fff4c6c624a03b75523a7352b5f9736f2d57e
Reviewed-on: https://review.typo3.org/26787
Reviewed-by: Jo Hasenau
Reviewed-by: Markus Klein
Tested-by: Markus Klein
Reviewed-by: Helmut Hummel
Reviewed-by: Oliver Hader
Tested-by: Oliver Hader