[TASK] Allow FrontendContentAdapter to be disabled
The FrontendContentAdapter transforms tt_content and pages records
within cObject to the format which they have been stored in in pre
6.x installations, so probably existing legacy TypoScript config
can deal with the same data.

This patchset introduces an AllConfiguration flag allowing to disable
that service. For new installations the service is disabled by default,
upgrading installations have the service enabled.

Resolves: #54979
Releases: 6.2
Change-Id: I254f5c170c3e6a9f5ee214c0fe0e7a58e080cf34
Reviewed-on: https://review.typo3.org/26807
Reviewed-by: Markus Klein
Tested-by: Markus Klein
Reviewed-by: Alexander Stehlik
Tested-by: Alexander Stehlik
Reviewed-by: Stefan Neufeind
Reviewed-by: Steffen Ritter
Tested-by: Steffen Ritter