[TASK] Skip two tests in LocalDriverTests in PHP 5.3
For most tests in the LocalDriverTest vfsStream is used to
emulate the filesystem. This needs PHP filesystem functions
to use the stream wrappers.

For touch() this has been enabled with PHP 5.4. The reversal
conclusion is, that this does not work with PHP 5.3 which
causes failing tests there.

As a result If the PHP version is 5.3 we skip the two tests.
This is a follow-up to https://review.typo3.org/#/c/26520/

Resolves: #54230
Releases: 6.2
Change-Id: I4e92927e1d77f825bab8583318a944a8916051d5
Reviewed-on: https://review.typo3.org/26845
Reviewed-by: Steffen Ritter
Tested-by: Steffen Ritter