[BUGFIX] Exception if directories are not readable
If a user has listing rights on a folder which contains
folders he does not have access to the user currently
will end up with Exceptions in folder trees, file lists
as well as element browsers.

Before FAL these folders just have been shown "locked".

This change introduces an "InaccessibleFolder" object
which will be created by the storage if a Folder is
requested which exists but there is no read access while
the user is allowed to list the contents of the parent
folder.

In addition the components have been adapted to use this
new information for a graceful behaviour.

WIP: Did not adapt the tests yet, since I need to do some
urgent customer work. Feel free to go on with that.

Resolves: #51512, #53854
Releases: 6.2
Change-Id: If5c234fe66d1930e4cc0cb16279d2f77874d82db
Reviewed-on: https://review.typo3.org/26905
Tested-by: Ingo Schmitt
Reviewed-by: Anja Leichsenring
Tested-by: Anja Leichsenring
Reviewed-by: Ingo Schmitt
Reviewed-by: Steffen Ritter
Tested-by: Steffen Ritter