[BUGFIX] Fix test loadValuesSetsStaticEnumConstants
If you run the test from the BE module of PHPUnit together with
all/other tests it fails. This is due to the fact that the
tests checks more than it actually should.

In particular the static $enumConstants array may already contain
information from other classes, therefore the test only has to
check for existane of information for the tested class.

The patch also corrects the missing \ in the namespace part
between Type and Fixture.

Releases: 6.2
Resolves: #55270
Change-Id: I1ea6546af8013f6c9075525c5bc6f3df579ccfcc
Reviewed-on: https://review.typo3.org/27020
Reviewed-by: Stefano Kowalke
Tested-by: Stefano Kowalke
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn