[TASK] Deleting currently opened folder in filelist results in error
When you have a folder opened in the filelist and then delete it
via the folder-tree the file-module shows an error after its reload.

This patch checks for existence of the folder before it is created.
If it is not existent it checks wether the parent folder exists and
is accessible. If, so the parent folder will be shown instead.

If the parent folder is not accessible the usual exception is thrown.

Releases: 6.2
Resolves: #54554
Change-Id: I7a21c76fd476ebce439489040f3b4c6786afd10a
Reviewed-on: https://review.typo3.org/27030
Reviewed-by: Marcin Sągol
Tested-by: Marcin Sągol
Reviewed-by: Marcel Wieser
Tested-by: Marcel Wieser
Reviewed-by: Steffen Ritter
Tested-by: Steffen Ritter