[TASK] Make files visible by default
For FAL, the concept of "hidden" files was introduced with the sysext
"metadata". The field is called "visible", its usage is to be
implemented by the plugin.

The default value for visible is FALSE, making it error-prone for users
uploading new files: They are hidden by default, requiring additional
work by the end-user (or the plugin/backend module doing the work) to
un-hide the files.

This commit changes the default of ``sys_file_metadata:visible`` to
``1``, making *new* files visible by default.

Change-Id: Id760ac92042e3235259ae78509187d3535cc25ca
Releases: 6.2
Resolves: #55410
Reviewed-on: https://review.typo3.org/27111
Reviewed-by: Andreas Wolf
Tested-by: Andreas Wolf