[TASK] Add functional test cases to DataHandler (live)
This change-set adds new functional tests for the DataHandler
to ensure the correct behaviour in the Live Workspace. The
difference to current functional tests is, that here a more
complete picture is checked in the database instead of only
particular ids and values.

The tests use a DataSet, which is basically only a CSV file
that hold the record values for required tables. These CSV
files can easily be modified by any spreadsheet application.
The "Scenario" DataSets are used to define the scenario each
test case is based on - the "Assertion" DataSets are used to
acutally assert that the correct processing was done in
the DataHandler.

Resolves: #54855
Releases: 6.2
Change-Id: I5d748cde04a70b9c158d09f9a0bd337ef809fd02
Reviewed-on: https://review.typo3.org/27188
Reviewed-by: Peter Kuehn
Tested-by: Peter Kuehn
Reviewed-by: Anja Leichsenring
Tested-by: Anja Leichsenring
Reviewed-by: Oliver Hader
Tested-by: Oliver Hader