[BUGFIX] Center icons vertically again in toolbar of page tree
The task #52497 introduced a regression. The default font-size 12px and
new default line-height of 18px set in body shifts the icons down. This
needs to be reset at least for toolbar icons in xtheme-t3skin.css.
Additionally removed the comment for styles no longer set in the
stylesheet.

Resolves: #55507
Related: #52497
Releases: 6.2
Change-Id: I02faf731ea41ad82338cbfc0b417f715ac0e13c1
Reviewed-on: http://review.typo3.org/35425
Tested-by: Benjamin Kott <info@bk2k.info>
Reviewed-by: Benjamin Kott <info@bk2k.info>
Reviewed-by: Benjamin Mack <benni@typo3.org>
Tested-by: Benjamin Mack <benni@typo3.org>