[BUGFIX] Installer rejects traling slash in link target
When having e.g. the link for ``typo3/`` pointing to
``typo3_src/typo3/``, the install tool complains because the trailing
slash is not expected. This slash is however perfectly ok, as there is
no difference between typo3_src/typo3 and typo3_src/typo3/ – both point
to the same "file" (i.e. node in the filesystem structure).

This patch strips trailing slashes from the actual link target before
doing the check, so no change to existing configuration is necessary.

Change-Id: Ibcb2118a2bb8fcf004470f1a949824576ad4d1ec
Resolves: #55551
Releases: 6.2
Reviewed-on: https://review.typo3.org/27229
Reviewed-by: Benjamin Mack
Reviewed-by: Christian Kuhn
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters
Reviewed-by: Stefan Neufeind
Tested-by: Stefan Neufeind